requirements.txt
aiohttp >= 3.9.3, < 4
python-dateutil >= 2.9.0, < 3
setuptools >= 69.1.1
urllib3 >= 1.25.11, < 3
opentelemetry-api >= 1.25.0, < 2
